backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
pyjwt
requests
